Recipe Detail Entity 수정

+++ src/main/java/com/kkkj/yorijori_be/Repository/Recipe/RecipeDetailRepository.java
@@ -3,5 +3,5 @@ package com.kkkj.yorijori_be.Repository.Recipe;
 import com.kkkj.yorijori_be.Entity.Recipe.RecipeDetailEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipeDetailRepository extends JpaRepository<RecipeDetailEntity, Integer> {
+public interface RecipeDetailRepository extends JpaRepository<RecipeDetailEntity, Long> {
 }
